initial scaffolding
Scaffold out the project and add a few CRD definitions with controllers
under the primaza.io group:
 - ClusterEnvironment
 - ServiceClass
 - ServiceClaim
 - ServiceBinding

Their controllers are left as no-ops; they will be fleshed out in future
stories.

Also define the following CRDs, without controllers:
 - RegisteredService
 - ServiceCatalog

All of the CRDs defined are left empty for the time being; their actual
structure will be added in the future.

Signed-off-by: Andy Sadler <ansadler@redhat.com>