continue testing even if another test fails
Currently, if an acceptance test fails, github will cancel all
incomplete acceptance tests.  This, unfortunately, can obscure the
effect of the change in question: since the tests don't run to
completion, we don't necessarily understand what tests are and are not
failing.

Change github's behavior to not cancel all test in the event of an early
test failure.

Signed-off-by: Andy Sadler <ansadler@redhat.com>