fix operator-sdk scaffolding
When we invoke operator-sdk to introduce new project scaffolding, we run
into the issue that it generates resources under the
`primaza.io.primaza.io` group name, not `primaza.io`.  This is caused by
a misconfiguration in the PROJECT file: we set both the domain and the
group field for each resource, which is not correct in this
circumstance.

To prevent problems from occuring in the future, remove the group field.

Signed-off-by: Andy Sadler <ansadler@redhat.com>