Allow secrets in mappings for SEDs
SED information could contain sensitive data, such as a password or
authentication token.  Such data should not be stored within a
registered service's resource itself.  When making registered services
with a service class, we should give users the option to indicate that
their data is secret.

This change introduces a `Secret` flag within a SED mapping, which
indicates whether that particular mapping item should be stored within
the registered service (false) or within an associated secret (true). By
default, this flag is true, since security features such as this should
be opt-out, not opt-in.

Signed-off-by: Andy Sadler <ansadler@redhat.com>